+++ pages/organisers-and-crew/jessica-white.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import Layout from '../../../components/Layout';
-import Header from '../../../components/Header';
-import theme from '../../../theme/theme';
+import Layout from '../../components/Layout';
+import Header from '../../components/Header';
+import theme from '../../theme/theme';
 import Head from 'next/head';
-import CrewBackButton from '../../../components/information/CrewBackButton';
+import BackToAboutButton from '../../components/information/BackToAboutButton';
 import { library } from '@fortawesome/fontawesome-svg-core';
 import {
   faInstagram,
@@ -14,77 +14,98 @@ import {
 
 library.add(faInstagram, faTwitter, faGithub, faLinkedin);
 
-export default function StevenPears() {
+export default function JessicaWhite() {
   return (
     <Layout>
       <Head>
-        <title>Steven Pears</title>
+        <title>Jessica White</title>
       </Head>
       <Header
-        title={'Steven Pears'}
+        title={'Jessica White'}
         banner="/static/banners/information/thencc.jpg"
       />
       <section>
-        <h1 tabIndex="0">Contracts and Finances Manager</h1>
+        <h1 tabIndex="0">Director/ Co-Founder / Co-Organiser</h1>
       </section>
       <section>
         <img
-          src="/static/organisers/steven.jpg"
-          alt="Picture of Steven Pears"
+          src="/static/organisers/jess.jpg"
+          alt="Picture of Jessica White"
           className="speakerpic"
         />
         <div>
           <ul>
-            <b>Pronouns: </b> He/Him
+            <b>Pronouns: </b> She/Her
           </ul>
         </div>
         <br />
       </section>
-
       <section>
-        <h2 tabIndex="0">About Steve</h2>
+        <h2 tabIndex="0">About Jessica</h2>
         <p tabIndex="0">
-          Steve (or StoiveyP) is a Technical Architect and active
-          member of the local community. He has volunteered at the
-          2019 and 2021 DDD East Midlands Conferences and provided a
-          lot of coffee to Jessica during the events. Steve is also a
-          Microsoft MVP and experienced public speaker.
+          Fully caffeinated, Jessica is more whirling dervish than
+          person. Professionally, she is a Technical Architect.
+          Outside of work she is a Microsoft MVP, Mentor, Organiser,
+          Gamer and avid reader.
+        </p>
+        <p tabIndex="0">
+          Jessica has a lot of event organisation experience. Having
+          founded{' '}
+          <a href="https://www.technottingham.com/wit-notts">
+            Women In Tech Nottingham
+          </a>
+          back in 2015, and running it for nearly 2 years, she has
+          previous experience with running an inclusive event. She
+          also co-organises the .NET Notts meetup in Nottingham.
         </p>
       </section>
 
       <section>
-        <h2 tabIndex="0">Steven's Responsibilities</h2>
-        <p tabIndex="0">
-          Will be working closely with Moreton to manage the events
-          contracts, budgets, payments and anything else financial.
-        </p>
+        <h2 tabIndex="0">Jessica's Responsibilities</h2>
+        <li tabIndex="0">Partner Agreements and outreach</li>
+        <li tabIndex="0">Scheduling</li>
+        <li tabIndex="0">Contracts</li>
+        <li tabIndex="0">Social Media</li>
+        <li tabIndex="0">Staffing</li>
+        <li tabIndex="0">Managing the Open Source Repositories</li>
+        <li tabIndex="0">Speaker Management</li>
       </section>
 
       <section>
         <h2 tabIndex="0">Links</h2>
         <li tabIndex="0">
           <a
+            href="mailto:jessica@dddeastmidlands.com"
             target="_blank"
             rel="noreferrer"
-            aria-label="Steven Pears: Twitter"
-            href="https://twitter.com/StevenPears"
+            aria-label="Jessica White: Email"
+          >
+            Email
+          </a>
+        </li>
+        <li tabIndex="0">
+          <a
+            href="https://twitter.com/JessPWhite"
+            target="_blank"
+            rel="noreferrer"
+            aria-label="Jessica White: Twitter"
           >
             Twitter
           </a>
         </li>
         <li tabIndex="0">
           <a
+            href="https://www.linkedin.com/in/jessica-white-67917883/"
             target="_blank"
             rel="noreferrer"
-            aria-label="Steven Pears: LinkedIn"
-            href="https://www.linkedin.com/in/stevenpears/"
+            aria-label="Jessica White: LinkedIn"
           >
             LinkedIn
           </a>
         </li>
       </section>
       <section>
-        <CrewBackButton />
+        <BackToAboutButton />
       </section>
 
       <style jsx>
